requirements.txt
requests>=2.31.0
pystray>=0.19.4
Pillow>=10.0.0 